Combing the code
Причесываем код
- Добавлена функция получения параметров объекта
-Функция получения типа объекта
-Код разбит на большее число файлов
(на старой версии студии компилится, у кого новая - исправьте ошибки, я пока обновляюсь)
