+++ ProgrammingBasics/ProgrammingBasics/Include/Model.h
@@ -1,7 +1,5 @@
 #include "Array.h"
-#include "Type.h"
-#include <cmath>
-#include "IDGenerator.h"
+#include "Primitives.h"
 
 #ifndef __MODEL
 #define __MODEL
@@ -14,92 +12,12 @@ private:
 	Array<Primitive*> data; //or another container
 	IDGenerator* idGen;
 public:
-	Model();
+	Model() {}
 	ID createObject(Type, Array<double>);
-	ID getNearest(double, double);
 	bool getNearest(double, double, ID&);
-	Type getObjType(ID);
-	Array<double> getObjSubs(ID); //?????
-	Array<double> getObjParam(ID);
+	bool getObjType(ID, Type&);
+	//Array<double> getObjSubs(ID); //?????
+	bool getObjParam(ID, Array<double>&);
 };
-
-// just pair of numbers
-class Vector2 {
-private:
-
-public:
-	Vector2(double, double);
-	Vector2();
-	static double Dot(Vector2, Vector2);
-	static double Cross(Vector2, Vector2);
-	double x;
-	double y;
-	Vector2 operator -(Vector2&);
-	Vector2 operator +(Vector2&);
-	Vector2 operator *(double);
-	double GetLength();
-};
-
-// parent of point, segment, arc
-class Primitive {
-private:
-	ID& id;
-	Type& type;
-public:
-	Primitive(ID, Type);
-	virtual double GetDistance(Vector2) = 0;
-	ID GetId();
-	Type GetType();
-};
-
-class Point : public Primitive {
-private:
-	Vector2 position;
-public:
-	Point(Vector2);
-	Point(double, double);
-	Point(Point&); //copying constructor
-	double GetDistance(Vector2) override;
-	Vector2 GetPosition() const;
-	void SetPosition(Vector2);
-	void SetPosition(double, double);
-};
-
-class Segment : public Primitive {
-private:
-	Point& point1;
-	Point& point2;
-public:
-	Segment(Vector2, Vector2);
-	Segment(double, double, double, double);
-	double GetDistance(Vector2) override;
-	ID GetPoint1_ID() const;
-	ID GetPoint2_ID() const;
-	Vector2 GetPoint1_pos() const;
-	Vector2 GetPoint2_pos() const;
-	void SetPoint1_pos(Vector2);
-	void SetPoint2_pos(Vector2);
-
-	/*functions for getting and setting coords*/
-};
-
-// piece of circle
-class Arc : public Primitive {
-private:
-	Point& point1;
-	Point& point2;
-	double angle; // from 0 to 2pi
-public: 
-	Arc(double, double, double, double, double); //size = sizeof(Arc);
-	Arc(Vector2, Vector2, double);
-	double GetDistance(Vector2) override;
-	Vector2 GetCenter() const;
-	/*Vector2 GetPoint1() const;
-	Vector2 GetPoint2() const;
-	other function for getting/setting
-	*/
-};
-
-
 
 #endif
