+++ ProgrammingBasics/ProgrammingBasics/Src/ID.cpp
@@ -1,0 +1,36 @@
+#include "IDGenerator.h"
+
+// ID FUNCTIONS
+
+ID::ID(unsigned long long h) {
+	hash = h;
+}
+
+bool ID::operator==(ID& item) {
+	return hash == item.hash;
+}
+
+bool ID::operator<(ID& item) {
+	return hash < item.hash;
+}
+
+ID::ID() {
+	hash = 0;
+}
+
+// IDGENERATOR FUNCTIONS
+
+ID IDGenerator::generateID() {
+	return ID(_lastGivenHash++);
+}
+
+IDGenerator* IDGenerator::getInstance() {
+	if (_instance == nullptr)
+		_instance = new IDGenerator;
+	return _instance;
+}
+
+// IDGENERATOR'S STATIC MEMBERS INITIALIZING
+
+unsigned long long IDGenerator::_lastGivenHash = 1;
+IDGenerator* IDGenerator::_instance = nullptr;
